+++ src/components/index.ts
@@ -1,4 +1,5 @@
 export * from './grid/grid.component';
+export * from './gridContainer/gridContainer.component';
 export * from './matrixGrid/matrixGrid.component';
 export * from './tableGridColumn/tableGridColumn.component';
 export * from './tableGridMetadataGatherer/tableGridMetadataGatherer.component';