+++ src/plugins/noDataRenderer/simple/simpleNoDataRenderer.component.ts
@@ -114,7 +114,8 @@ export class SimpleNoDataRendererSAComponent implements SimpleNoDataRenderer, Gr
             this.dataLoader = dataLoader;
         }
         
-        //TODO: tests whether 
+        //TODO: tests whether creating new computet destroyes existing one
+        //TODO: probably only need when force, or data loader changes
         this.text = computed<string>(() =>
         {
             if(!this.dataLoader)